textshaping-package {textshaping}	R Documentation

textshaping: Bindings to the 'HarfBuzz' and 'Fribidi' Libraries for Text Shaping

Description

Provides access to the text shaping functionality in the 'HarfBuzz' library and the bidirectional algorithm in the 'Fribidi' library. 'textshaping' is a low-level utility package mainly for graphic devices that expands upon the font tool-set provided by the 'systemfonts' package.

Author(s)

Maintainer: Thomas Lin Pedersen thomas.pedersen@posit.co (ORCID)

Other contributors:

Posit, PBC [copyright holder, funder]

See Also

Useful links:

https://github.com/r-lib/textshaping

Report bugs at https://github.com/r-lib/textshaping/issues

[Package textshaping version 0.3.7 Index]